Fix typo from AllProvidedTypeShouldBeInterface -> AllProvidedTypesShouldBeInterface
