+++ src/Pipelines/Builder/Validators/AllProvidedTypesShouldBeInterface.cs
@@ -2,7 +2,7 @@ using Pipelines.Exceptions;
 
 namespace Pipelines.Builder.Validators;
 
-internal static class AllProvidedTypeShouldBeInterface
+internal static class AllProvidedTypesShouldBeInterface
 {
     internal static void Validate(params Type[] typesToValidate)
     {
