--- tests/Pipelines.Tests/Builder/Validators/AllProvidedTypeShouldBeInterfaces/Types/ExampleEnum.cs
@@ -1,6 +1,0 @@
-namespace Pipelines.Tests.Builder.Validators.AllProvidedTypeShouldBeInterfaces.Types;
-
-public enum ExampleEnum
-{
-    
-}
