--- tests/Pipelines.Tests/Builder/Validators/AllProvidedTypeShouldBeInterfaces/Types/ExampleRecord.cs
@@ -1,6 +1,0 @@
-namespace Pipelines.Tests.Builder.Validators.AllProvidedTypeShouldBeInterfaces.Types;
-
-public record ExampleRecord
-{
-    
-}
