--- tests/Pipelines.Tests/Builder/Validators/AllProvidedTypeShouldBeInterfaces/Types/ExampleStruct.cs
@@ -1,6 +1,0 @@
-namespace Pipelines.Tests.Builder.Validators.AllProvidedTypeShouldBeInterfaces.Types;
-
-public struct ExampleStruct
-{
-    
-}
