+++ tests/Pipelines.Tests/Builder/Validators/AllProvidedTypesShouldBeInterfaces/Types/ExampleRecord.cs
@@ -1,0 +1,6 @@
+namespace Pipelines.Tests.Builder.Validators.AllProvidedTypesShouldBeInterfaces.Types;
+
+public record ExampleRecord
+{
+    
+}
