+++ tests/Pipelines.Tests/Builder/Validators/AllProvidedTypesShouldBeInterfaces/Types/ExampleStruct.cs
@@ -1,0 +1,6 @@
+namespace Pipelines.Tests.Builder.Validators.AllProvidedTypesShouldBeInterfaces.Types;
+
+public struct ExampleStruct
+{
+    
+}
